add dry run option

+++ main.py
@@ -52,25 +52,37 @@ def upload():
     return("Uploaded Files")
 
 @app.route('/process', methods=['GET'])
-def process_uploads(media_type = None):
+def process_uploads(media_type=None):
     # connect google drive / plex library folder
     if media_type == None:
         media_type = request.args['media_type']
-    print(media_type)
+    
+    dry_run = bool(request.args['dry_run'])
     #media_type = 'movie' #request.data['media_type']
     UPLOADS = os.path.join(app.config['UPLOAD_FOLDER'], media_type)
 
     if media_type == 'movie':
         # move uploaded file to plex/movies
         dirs = os.listdir(UPLOADS)
+        errs = []
         for dir in dirs:
-            shutil.move(os.path.join(UPLOADS, dir), app.config['PLEX_MOVIE_FOLDER'])
+            try:
+                shutil.move(os.path.join(UPLOADS, dir), app.config['PLEX_MOVIE_FOLDER'])
+            except shutil.Error as e:
+                dirs.remove(dir) 
+                errs.append(str(e))
+
         
-        return(json.dumps(dirs))
+        return {
+            "files": dirs,
+            "errors": errs,
+            "plex_url": "https://app.plex.tv/desktop#!/media/d391e47fb3d0e2f458f498e7a5dbf6b20f3ec2b4/com.plexapp.plugins.library?key=%2Flibrary%2Fsections%2F1%2Fall%3Fsort%3DaddedAt%3Adesc&title=Recently%20Added%20in%20Films&context=source%3Ahub.movie.recentlyadded&pageType=list&source=1"
+            }
+
     elif media_type == 'tv':
         # Detect tv show using tvnamer
         try:
-            episodes = detect_shows(UPLOADS)
+            episodes = detect_shows(UPLOADS, dry_run)
             episodes_response = [episode.generateFilename() for episode in episodes]
         except NoValidFilesFoundError:
             return("No valid files were supplied")
@@ -80,33 +92,36 @@ def process_uploads(media_type = None):
             return(errormsg)
 
         # move uploaded file to plex folder
-        for episode in episodes:
-            if episode.seriesname:
-                dest_dir = os.path.join(
-                    app.config['PLEX_TV_FOLDER'],
-                    episode.seriesname.lower(),
-                    "season_{:02d}".format(episode.seasonnumber)
-                    )
+        if not dry_run:
+            for episode in episodes:
+                if episode.seriesname:
+                    dest_dir = os.path.join(
+                        app.config['PLEX_TV_FOLDER'],
+                        episode.seriesname.lower(),
+                        "season_{:02d}".format(episode.seasonnumber)
+                        )
 
-                os.makedirs(dest_dir, exist_ok=True)
+                    os.makedirs(dest_dir, exist_ok=True)
 
-                parent_dir = "/".join(episode.fullpath.split("/")[:-1])
-                filename = episode.generateFilename()
-                fullpath = os.path.join(parent_dir, filename)
+                    parent_dir = "/".join(episode.fullpath.split("/")[:-1])
+                    filename = episode.generateFilename()
+                    fullpath = os.path.join(parent_dir, filename)
 
-                shutil.move(fullpath, os.path.join(dest_dir, filename))
-                # delete season directory if empty
-                if not os.listdir(parent_dir):
-                    shutil.rmtree(parent_dir)
+                    shutil.move(fullpath, os.path.join(dest_dir, filename))
+                    # delete season directory if empty
+                    if not os.listdir(parent_dir):
+                        shutil.rmtree(parent_dir)
 
         # delete series directory if empty
-        print(os.listdir(UPLOADS))
         for dir in os.listdir(UPLOADS):
-            print(dir)
-            if not os.listdir(os.path.join(UPLOADS, dir)):
+            if os.path.isdir(dir) and not os.listdir(os.path.join(UPLOADS, dir)):
                 shutil.rmtree(dir)
         
-        return json.dumps(episodes_response)
+        return {
+            "files": episodes_response,
+            "errors": [],
+            "plex_url": "https://app.plex.tv/desktop#!/media/d391e47fb3d0e2f458f498e7a5dbf6b20f3ec2b4/com.plexapp.plugins.library?key=%2Fhubs%2Fhome%2FrecentlyAdded%3Ftype%3D2%26sectionID%3D2&title=Recently%20Added%20in%20TV%20programmes&context=source%3Ahub.tv.recentlyadded&pageType=list&source=2"
+            }
 
 
     else:
